Dose-response model function for listeriosis

The DR() function provides the marginal probability of invasive listeriosis in a given population for a given Dose in CFU using the JEMRA, the Pouillot, the Fritsch, the EFSA dose-response models or the model developed within this project (EFSAMV,EFSAV,EFSALV) (see References).

Usage

DR(
  Dose,
  model = "JEMRA",
  population = 1,
  Poisson = FALSE,
  method = "integrate",
  ...
)

Arguments

Dose

average dose in CFU/serving assuming a Poisson distribution from serving to serving (scalar or vector)

model

either JEMRA, Pouillot, Fritsch, EFSA, EFSAMV,EFSAV or EFSALV

population

considered population (scalar or vector) (see below)

Poisson

if TRUE, assume that Dose is the mean of a Poisson distribution. (actual LogNormal Poisson). If FALSE (default), assume that Dose is the actual number of bacteria.

method

either integrate or cubature to specify the integration method.

...

further arguments to pass to the DRLogNormPoisson() function

Value

A vector of size Dose (if population is a scalar) or a matrix of dimension (length of the Dose vector times length of the population vector)

Details

For the Pouillot, Fritsch, EFSA EFSAMV,EFSAV or EFSALV model, it integrates a log10Normal Poisson distribution Pouillot2015;textualDRLm. or a log10Normal-binomial distribution. For the JEMRA model an exponential dose-response model or a binomial model is used.

This function is slow. Use DRQuick() in production.

Model	Population	Characteristics

JEMRA	0	Marginal over subpopulations

JEMRA	1	Healthy population

JEMRA	2	Increased susceptibility

Pouillot	0	Marginal over subpopulations

Pouillot	1	Less than 65 years old

Pouillot	2	More than 65 years old

Pouillot	3	Pregnancy

Pouillot	4	Nonhematological Cancer

Pouillot	5	Hematological cancer

Pouillot	6	Renal or Liver failure

Pouillot	7	Solid organ transplant

Pouillot	8	Inflammatory diseases

Pouillot	9	HIV/AIDS

Pouillot	10	Diabetes

Pouillot	11	Hear diseases

Fritsch	0	Marginal over subpopulations

Fritsch	1	Highly virulent

Fritsch	2	Medium virulent

Fritsch	3	Hypovirulent

EFSA-EFSALV-EFSAV-EFSAMV	0	Marginal over subpopulations

EFSA-EFSALV-EFSAV-EFSAMV	1	Female 1-4 yo

EFSA-EFSALV-EFSAV-EFSAMV	2	Male 1-4 yo

EFSA-EFSALV-EFSAV-EFSAMV	3	Female 5-14 yo

EFSA-EFSALV-EFSAV-EFSAMV	4	Male 5-14 yo

EFSA-EFSALV-EFSAV-EFSAMV	5	Female 15-24 yo

EFSA-EFSALV-EFSAV-EFSAMV	6	Male 15-24 yo

EFSA-EFSALV-EFSAV-EFSAMV	7	Female 25-44 yo

EFSA-EFSALV-EFSAV-EFSAMV	8	Male 25-44 yo

EFSA-EFSALV-EFSAV-EFSAMV	9	Female 45-64 yo

EFSA-EFSALV-EFSAV-EFSAMV	10	Male 45-64 yo

EFSA-EFSALV-EFSAV-EFSAMV	11	Female 65-74 yo

EFSA-EFSALV-EFSAV-EFSAMV	12	Male 65-74 yo

EFSA-EFSALV-EFSAV-EFSAMV	13	Female >75 yo

EFSA-EFSALV-EFSAV-EFSAMV	14	Male >75 yo

See the parameters in the JEMRA report.

method = "cubature" will use the hcubature function that is much slower but guarantees a tolerance of \(1E-5\).

References

EFSA2018DRLm

FAO-WHO2004DRLm

Fritsch2018DRLm

Pouillot2015DRLm

Author

Regis Pouillot

Examples

DR(5:10, "Pouillot", 1:11)
#>      Less than 65 years old, no known underlying condition
#> [1,]                                          4.077694e-11
#> [2,]                                          4.893216e-11
#> [3,]                                          5.708732e-11
#> [4,]                                          6.524243e-11
#> [5,]                                          7.339749e-11
#> [6,]                                          8.155250e-11
#>      More than 65 years old, no known underlying condition    Pregnancy
#> [1,]                                          7.768332e-10 1.046297e-08
#> [2,]                                          9.321587e-10 1.255169e-08
#> [3,]                                          1.087472e-09 1.463939e-08
#> [4,]                                          1.242774e-09 1.672614e-08
#> [5,]                                          1.398066e-09 1.881197e-08
#> [6,]                                          1.553346e-09 2.089694e-08
#>      Nonhematological Cancer Hematological cancer Renal or Liver failure
#> [1,]            4.074294e-09         4.989462e-08           1.443598e-08
#> [2,]            4.888381e-09         5.982261e-08           1.731652e-08
#> [3,]            5.702257e-09         6.973809e-08           2.019534e-08
#> [4,]            6.515934e-09         7.964192e-08           2.307254e-08
#> [5,]            7.329420e-09         8.953481e-08           2.594821e-08
#> [6,]            8.142723e-09         9.941735e-08           2.882241e-08
#>      Solid organ transplant Inflammatory diseases     HIV/AIDS     Diabetes
#> [1,]           1.619420e-08          4.365476e-09 3.389244e-09 3.893770e-10
#> [2,]           1.942499e-08          5.237701e-09 4.066531e-09 4.672406e-10
#> [3,]           2.265369e-08          6.109690e-09 4.743665e-09 5.451009e-10
#> [4,]           2.588045e-08          6.981455e-09 5.420653e-09 6.229577e-10
#> [5,]           2.910535e-08          7.853005e-09 6.097502e-09 7.008114e-10
#> [6,]           3.232850e-08          8.724350e-09 6.774215e-09 7.786619e-10
#>      Hear diseases
#> [1,]  2.632604e-10
#> [2,]  3.159067e-10
#> [3,]  3.685514e-10
#> [4,]  4.211944e-10
#> [5,]  4.738358e-10
#> [6,]  5.264757e-10
DR(5:10, "Pouillot", 0)
#>              [,1]
#> [1,] 6.210723e-10
#> [2,] 7.450437e-10
#> [3,] 8.689530e-10
#> [4,] 9.928042e-10
#> [5,] 1.116600e-09
#> [6,] 1.240344e-09