Marginal probability of invasive listeriosis over r

The function DRLogNormPoisson() provides the marginal probability of invasive listeriosis in a given population for a given Dose in CFU. this function is not vectorized.

Usage

DRLogNormPoisson(
  Dose,
  meanlog10,
  sdlog10,
  Poisson = FALSE,
  low = -Inf,
  up = Inf,
  silent = TRUE,
  tol = 1e-20,
  method = "integrate",
  ...
)

Arguments

Dose

(CFU/serving) Dose (scalar or vector). It should be integers if Poisson is FALSE.

meanlog10

the meanlog10 parameter of the distribution of r (parameter of the exponential model).

sdlog10

the sdlog10 parameter of the distribution of r (parameter of the exponential model).

Poisson

if TRUE, assume that Dose is the mean of a Poisson distribution. (actual LogNormal Poisson). If FALSE (default), assume that Dose is the actual number of bacteria.

low

lower value for the integration.

up

upper value for the integration.

silent

silent the error-try function.

tol

relative tolerance. Note: for method = "cubature", the tolerance will be set to \(1E-05\).

method

either "integrate" (default) or "cubature" to specify the integration method.

...

further arguments to pass to the integrate function.

Value

Probability of invasive listeriosis integrated over r.

Details

The function evaluates $$\int_{low}^{inf} \Phi(x, mulog_{10}, sdlog_{10})\cdot(1-e^{(-Dose \cdot 10^{r})}) dr$$ using the integrate function, with a relative tolerance equals to tol if Poisson is TRUE. If Poisson is FALSE, it evaluates $$\int_{low}^{inf} \Phi(x, mulog_{10}, sdlog_{10})\cdot(1-(1-10^{r})^{Dose}) dr$$.

For method = "cubature", the tolerance will be set to \(1E-5\). method = "cubature" will use the \link[cubature]{hcubature} function that is much slower but guarantees a tolerance of \(1E-5\).

Note

This function is used by the DR() function, a wrapper of DRLogNormPoisson(). For a quick, vectorized version of it, use DRQuick().

References

Pouillot2015DRLm

Author

Régis Pouillot

Examples

DRLogNormPoisson(2, -14.11, 1.62,low=-Inf,up=Inf)
#> [1] 1.631096e-11
DRLogNormPoisson(2, -14.11, 1.62, Poisson=TRUE)
#> [1] 1.631093e-11