Dose-response model function for listeriosis - quick version

This function provides the marginal probability of invasive listeriosis in a given population for a given Dose in CFU using the JEMRA, the Pouillot, the Fritsch or the EFSA dose-response models or the model developed within this project (EFSAMV,EFSAV,EFSALV) (see References).

Usage

DRQuick(Dose, model = "JEMRA", population = 1, Poisson = FALSE)

Arguments

Dose

(CFU/serving) Dose (scalar or vector).

model

either JEMRA, Pouillot, Fritsch, EFSA, EFSAMV,EFSAV or EFSALV

population

considered population (scalar or vector).

Poisson

if TRUE, assume that Dose is the mean of a Poisson distribution. (actual LogNormal Poisson). If FALSE (default), assume that Dose is the actual number of bacteria.

Value

A vector of size Dose (if population is a scalar) or a matrix of dimension (length of the Dose vector x length of the population vector)

Details

Model	Population	Characteristics

JEMRA	1	Healthy population

JEMRA	2	Increased susceptibility

Pouillot	1	Less than 65 years old

Pouillot	2	More than 65 years old

Pouillot	3	Pregnancy

Pouillot	4	Nonhematological Cancer

Pouillot	5	Hematological cancer

Pouillot	6	Renal or Liver failure

Pouillot	7	Solid organ transplant

Pouillot	8	Inflammatory diseases

Pouillot	9	HIV/AIDS

Pouillot	10	Diabetes

Pouillot	11	Hear diseases

Fritsch	1	Highly virulent

Fritsch	2	Medium virulent

Fritsch	3	Hypovirulent

EFSA-EFSALV-EFSAV-EFSAMV	1	Female 1-4 yo

EFSA-EFSALV-EFSAV-EFSAMV	2	Male 1-4 yo

EFSA-EFSALV-EFSAV-EFSAMV	3	Female 5-14 yo

EFSA-EFSALV-EFSAV-EFSAMV	4	Male 5-14 yo

EFSA-EFSALV-EFSAV-EFSAMV	5	Female 15-24 yo

EFSA-EFSALV-EFSAV-EFSAMV	6	Male 15-24 yo

EFSA-EFSALV-EFSAV-EFSAMV	7	Female 25-44 yo

EFSA-EFSALV-EFSAV-EFSAMV	8	Male 25-44 yo

EFSA-EFSALV-EFSAV-EFSAMV	9	Female 45-64 yo

EFSA-EFSALV-EFSAV-EFSAMV	10	Male 45-64 yo

EFSA-EFSALV-EFSAV-EFSAMV	11	Female 65-74 yo

EFSA-EFSALV-EFSAV-EFSAMV	12	Male 65-74 yo

EFSA-EFSALV-EFSAV-EFSAMV	13	Female >75 yo

EFSA-EFSALV-EFSAV-EFSAMV	14	Male >75 yo

See the parameters in the JEMRA report.

Note

This function uses (for all model but JEMRA) a linear approximation (approxfun) from the exact DR() model evaluated on \(Dose = c(0,10^{seq(-5,12,length=1701)})\) (if Poisson=TRUE) or \(c(0,10^{seq(0,12,length=2000)})\) (if Poisson=FALSE). Any Dose lower or higher than these ranges will lead to NA.

References

EFSA2018DRLm

FAO-WHO2004DRLm

Fritsch2018DRLm

Pouillot2015DRLm

See also

Author

Regis Pouillot

Examples

# Compare DR and DRQuick
cbind(DR(1:10, model="Pouillot", population = 5), 
      DRQuick(1:10,  model="Pouillot", population = 5))
#>       Hematological cancer Hematological cancer
#>  [1,]         1.002798e-08         1.002798e-08
#>  [2,]         2.002244e-08         2.002244e-08
#>  [3,]         2.999631e-08         2.999631e-08
#>  [4,]         3.995300e-08         3.995300e-08
#>  [5,]         4.989462e-08         4.989462e-08
#>  [6,]         5.982261e-08         5.982261e-08
#>  [7,]         6.973809e-08         6.973809e-08
#>  [8,]         7.964192e-08         7.964192e-08
#>  [9,]         8.953481e-08         8.953481e-08
#> [10,]         9.941735e-08         9.941735e-08
DRQuick(1:10,  model="Pouillot", population = 2)
#>       More than 65 years old, no known underlying condition
#>  [1,]                                          1.553994e-10
#>  [2,]                                          3.107796e-10
#>  [3,]                                          4.661445e-10
#>  [4,]                                          6.214954e-10
#>  [5,]                                          7.768332e-10
#>  [6,]                                          9.321586e-10
#>  [7,]                                          1.087472e-09
#>  [8,]                                          1.242774e-09
#>  [9,]                                          1.398066e-09
#> [10,]                                          1.553346e-09